mobile-service/requirements.txt
Django>=4.2,<5.0
psycopg2-binary
djangorestframework
PyJWT
django-cors-headers
kafka-python==2.0.2
